make sure that just one thread at a time can crate new bpartner-stats
..and only on app.
to avoid the error "ERROR: duplicate key value violates unique constraint "c_bpartner_product_stats_uq""
