+++ backend/de.metas.business/src/main/java/de/metas/bpartner/product/stats/BPartnerProductStatsEventListener.java
@@ -1,11 +1,5 @@
 package de.metas.bpartner.product.stats;
 
-import javax.annotation.PostConstruct;
-
-import org.slf4j.Logger;
-import org.springframework.context.annotation.Profile;
-import org.springframework.stereotype.Component;
-
 import de.metas.Profiles;
 import de.metas.event.Event;
 import de.metas.event.IEventBus;
@@ -13,6 +7,12 @@ import de.metas.event.IEventBusFactory;
 import de.metas.event.IEventListener;
 import de.metas.logging.LogManager;
 import lombok.NonNull;
+import org.slf4j.Logger;
+import org.springframework.context.annotation.Profile;
+import org.springframework.stereotype.Component;
+
+import javax.annotation.PostConstruct;
+import java.util.concurrent.locks.ReentrantLock;
 
 /*
  * #%L
@@ -42,15 +42,24 @@ public class BPartnerProductStatsEventListener implements IEventListener
 {
 	private static final Logger logger = LogManager.getLogger(BPartnerProductStatsEventListener.class);
 
+	/**
+	 * Introduce this lock to get rid of
+	 * <pre>
+	 *     ERROR: duplicate key value violates unique constraint "c_bpartner_product_stats_uq"
+	 * </pre>
+	 * ..I suspect that the reason for the problem is that this listener is invoked concurrently, e.g. from async processors
+	 */
+	private final ReentrantLock statsRepoAccessLock = new ReentrantLock();
+
 	private final IEventBusFactory eventBusRegistry;
-	private final BPartnerProductStatsService statsService;
+	private final BPartnerProductStatsEventHandler statsEventHandler;
 
 	public BPartnerProductStatsEventListener(
 			@NonNull final IEventBusFactory eventBusRegistry,
-			@NonNull final BPartnerProductStatsService statsService)
+			@NonNull final BPartnerProductStatsEventHandler statsEventHandler)
 	{
 		this.eventBusRegistry = eventBusRegistry;
-		this.statsService = statsService;
+		this.statsEventHandler = statsEventHandler;
 	}
 
 	@PostConstruct
@@ -64,17 +73,25 @@ public class BPartnerProductStatsEventListener implements IEventListener
 	public void onEvent(final IEventBus eventBus, final Event event)
 	{
 		final String topicName = eventBus.getTopicName();
-		if (BPartnerProductStatsEventSender.TOPIC_InOut.getName().equals(topicName))
+		statsRepoAccessLock.lock();
+		try
 		{
-			statsService.handleInOutChangedEvent(BPartnerProductStatsEventSender.extractInOutChangedEvent(event));
+			if (BPartnerProductStatsEventSender.TOPIC_InOut.getName().equals(topicName))
+			{
+				statsEventHandler.handleInOutChangedEvent(BPartnerProductStatsEventSender.extractInOutChangedEvent(event));
+			}
+			else if (BPartnerProductStatsEventSender.TOPIC_Invoice.getName().equals(topicName))
+			{
+				statsEventHandler.handleInvoiceChangedEvent(BPartnerProductStatsEventSender.extractInvoiceChangedEvent(event));
+			}
+			else
+			{
+				logger.warn("Ignore unknown event {} got on topic={}", event, topicName);
+			}
 		}
-		else if (BPartnerProductStatsEventSender.TOPIC_Invoice.getName().equals(topicName))
+		finally
 		{
-			statsService.handleInvoiceChangedEvent(BPartnerProductStatsEventSender.extractInvoiceChangedEvent(event));
-		}
-		else
-		{
-			logger.warn("Ignore unknow event {} got on topic={}", event, topicName);
+			statsRepoAccessLock.unlock();
 		}
 	}
 }
